backend/requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# AI & ML
google-genai>=1.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0

# OCR & Image Processing
pytesseract>=0.3.10
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
celery>=5.3.6
redis>=5.0.1

# Production
gunicorn>=21.0.0
whitenoise>=6.6.0

# Development
pytest>=8.0.0
pytest-django>=4.8.0
black>=24.1.0
flake8>=7.0.0
